--- src/main/java/org/example/printForm/PrintFromFile.java
@@ -1,17 +1,0 @@
-package org.example.printForm;
-
-import static org.example.printForm.PrintDefault.*;
-import static org.example.source.XlsxReadUtils.*;
-
-public class PrintFromFile {
-    private PrintFromFile(){}
-
-    public static void printFromFileStudent(String link) {
-        printStudents(loadStudentsFromFile(link));
-    };
-
-    public static void printFromFileUniversity(String link) {
-        printUniver(loadUniversitiesFromFile(link));
-    }
-}
-
